+++ src/common/PlayerState.cxx
@@ -97,10 +97,10 @@ void*	PlayerState::pack(void* buf, uint16_t& code)
     // put the angle between -M_PI and +M_PI
     float angle = fmodf (azimuth, (float)M_PI * 2.0f);
     if (angle > M_PI) {
-      angle -= (M_PI * 2.0f);
+      angle -= (float)(M_PI * 2.0);
     }
     else if (angle < -M_PI) {
-      angle += (M_PI * 2.0f);
+      angle += (float)(M_PI * 2.0);
     }
     aziShort = (int16_t) ((angle * smallScale) / M_PI);
     angVelShort = (int16_t) ((angVel * smallScale) / smallMaxAngVel);
@@ -170,7 +170,7 @@ void*	PlayerState::unpack(void* buf, uint16_t code)
       pos[i] = ((float)posShort[i] * smallMaxDist) / smallScale;
       velocity[i] = ((float)velShort[i] * smallMaxVel) / smallScale;
     }
-    azimuth = ((float)aziShort * M_PI) / smallScale;
+    azimuth = (float)((aziShort * M_PI) / smallScale);
     angVel = ((float)angVelShort * smallMaxAngVel) / smallScale;
   }
